fix(build): 清理所有编译警告（clean build 零警告）
清理 MSVC + MinGW 的全部编译警告：

1. D9025 /MD→/MT：用 CMAKE_MSVC_RUNTIME_LIBRARY 替代 /MT 命令行选项
2. D9025 /O2→/O1：清空 CMAKE_C_FLAGS_RELEASE 默认 /O2，由各 target 自行指定
3. C4319 ~零扩展：~(DWORD)(align-1) 改为 ~(size_t)(align-1)（builder.c + builder_reflective.c）
4. C4996 strdup：改用 _strdup（MSVC 弃用 POSIX 名）
5. MinGW __winlock_sfence：移除中间 static 函数，直接用内联汇编
6. MinGW xtea_encrypt_buf unused：GCC 分支 WINLOCK_XTEA_FN 加 unused 属性

验证：build.ps1 -Clean 零警告，e2e 测试 36/36 全部通过

+++ packer/reflective/builder_reflective.c
@@ -768,12 +768,14 @@ int main(int argc, char* argv[]) {
     }
     if (out_size > last_raw_end) last_raw_end = (DWORD)out_size;
 
-    /* 12. 计算新 .payload 节的位置（基于最终的节布局） */
-    DWORD new_va           = (last_va_end + s_sec_align - 1) & ~(s_sec_align - 1);
+    /* 12. 计算新 .payload 节的位置（基于最终的节布局）
+     * 注意：~(align-1) 必须用 size_t 计算，避免 DWORD 的 ~ 高 32 位为零
+     * 导致 C4319 警告和对齐计算错误 */
+    DWORD new_va           = (DWORD)((last_va_end + s_sec_align - 1) & ~(size_t)(s_sec_align - 1));
     DWORD new_vsize        = (DWORD)total_payload_size;
-    DWORD new_vsize_aligned= (new_vsize + s_sec_align - 1) & ~(s_sec_align - 1);
-    DWORD new_raw_off      = (last_raw_end + s_file_align - 1) & ~(s_file_align - 1);
-    DWORD new_raw_size     = (DWORD)((total_payload_size + s_file_align - 1) & ~(s_file_align - 1));
+    DWORD new_vsize_aligned= (DWORD)((new_vsize + s_sec_align - 1) & ~(size_t)(s_sec_align - 1));
+    DWORD new_raw_off      = (DWORD)((last_raw_end + s_file_align - 1) & ~(size_t)(s_file_align - 1));
+    DWORD new_raw_size     = (DWORD)((total_payload_size + s_file_align - 1) & ~(size_t)(s_file_align - 1));
 
     printf("[+] New .payload section: VA=0x%lx VSize=0x%lx RawOff=0x%lx RawSize=0x%lx\n",
            (unsigned long)new_va, (unsigned long)new_vsize,
